style: interpolate harmony color
