+++ packages/color/package.json
@@ -10,7 +10,6 @@
     "./utils": "./src/utils.ts",
     "./convert": "./src/lib/convert.ts",
     "./gamut": "./src/lib/gamut.ts",
-    "./harmony": "./src/lib/harmony.ts",
     "./interpolate": "./src/lib/interpolate.ts",
     "./params": "./src/lib/params.ts",
     "./parse": "./src/lib/parse.ts",
